Moved lib/thing/index.js files to lib/thing.js
no issue

- cleans up directory structure and increases naming consistency
